Team Rankings and fix scores not working after 1st game
Added Team Rankings to Scores page. Still needs some formatting tweaks but works.
Fixed issue where playerInfo,isPlayerDebriefed was not set false when starting a new game, so scoring never worked on any game after the 1st game.
Tweaking Rehost CountdownTime. Now gameInfo.countdownTime, but also minimum of 10 seconds as some taggers are slow to rehost,

Lots of C++17 modifications and Qt deprecated code fixes.

+++ Game.h
@@ -152,7 +152,7 @@ private:
 	QString	currentGameFileName;
 
 	int     GameType;
-    int     GameID;
+    uint8_t GameID;
     int     GameLength;
     int     NumberOfPlayers;
     int     NumberOfTeams;
